breakout getWalletData into utils and add coverage
