+++ test/test.js
@@ -1,6 +1,7 @@
 const assert = require('assert');
 const nock = require('nock');
 const { describe, it } = require("mocha");
+const { hasAvailableDispensers, getWalletData } = require('../scripts/dispensers/utils.js');
 
 // Mock xchain.io/api/dispensers/{walletId} response
 nock('https://xchain.io')
@@ -57,6 +58,24 @@ describe('Utils', function () {
 
 		it('should return false because no value was passed in (handled error)', function () {
 			assert.equal(hasAvailableDispensers(), false);
+		});
+	});
+	describe('getWalletData()', function () {
+		it('should return wallet data', function () {
+			getWalletData('1').then((response) => {
+				assert.equal(response.id, 1);
+				assert.equal(response.total, 1);
+				assert.equal(response.data[0].asset, 'Asset Name');
+				assert.equal(response.data[0].escrow_quantity, '4');
+				assert.equal(response.data[0].give_quantity, '1');
+				assert.equal(response.data[0].give_remaining, '4');
+				assert.equal(response.data[0].status, '10');
+			});
+		});
+		it('should error if wallet id is not present', function () {
+			const borkedData = getWalletData()
+			assert.equal(borkedData.name, 'Error');
+			assert.equal(borkedData.message, 'Wallet ID is not a string');
 		});
 	});
 });
